src/test/javascript/spec/app/entities/ward/ward.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { WardService } from 'app/entities/ward/ward.service';
import { IWard, Ward } from 'app/shared/model/ward.model';
import { ClassType } from 'app/shared/model/enumerations/class-type.model';
import { Location } from 'app/shared/model/enumerations/location.model';

describe('Service Tests', () => {
  describe('Ward Service', () => {
    let injector: TestBed;
    let service: WardService;
    let httpMock: HttpTestingController;
    let elemDefault: IWard;
    let expectedResult: IWard | IWard[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(WardService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Ward(0, 'AAAAAAA', 'AAAAAAA', ClassType.A, Location.A1);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Ward', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Ward()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Ward', () => {
        const returnedFromService = Object.assign(
          {
            wardReferenceId: 'BBBBBB',
            wardName: 'BBBBBB',
            wardClassType: 'BBBBBB',
            wardLocation: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Ward', () => {
        const returnedFromService = Object.assign(
          {
            wardReferenceId: 'BBBBBB',
            wardName: 'BBBBBB',
            wardClassType: 'BBBBBB',
            wardLocation: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Ward', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
